Correct prefix parsing
Prefixes were being resolved right before the operator operation. This made it impossible to tell if the Symbol came from parentheses or if it was the first object with the complex symbol. To correct this the time of the prefix resolution was changed to the time when the Symbol was being inserted on the output stack therefore ensuring that the prefix always belonged to the symbol being placed on the stack.