fix: a/x/y-a/(x*y) not simplifying & scientific #
Scientific numbers were not being recognized correctly and a/x/y-a/(x*y)  was not simplifying to zero. To fix this issue scientific numbers are first removed with a place holder and then replaced back right before conversion to a symbol. To fix a/x/y-a/(x*y) all powers are kept positive during multiplication.