Correction to sort order for long division
The sort order in multivariate polynomial long division would result in the incorrect selection of the largest term. This could be seen in  divide(x^2*y+4*y+x^5+11*x^3+28*x,x^3+7*x+y) for example which would cause the power in the dividend to grow and therefore divergence from the answer.