Non-linear solve working.
The function works and was sped up a bit but still has some flaws. The first flaw is the fact that depending on how far away we are from the roots, it does not converge. The second is that depending on our start point, we can get a different set of roots.
Consider nerdamer.solveEquations(['3*x^2+y-4*z=5', '4*y^3+x+z^2=-4', 'x-y-6*z/x=-5']);
When starting at 0.01, the solution set which is returned is x=-0.95998068960715, y=-0.97287943854958, z=-0.80204766632343 which is not the solution set I was looking for. Changing the start to 30 the function yields x=3, y=-2, z=5, which is the solution set I was looking for. Improvements are warranted but the onus lies on the user to poke and prod a bit.